Changes to handle calls with a “/“ in the call sign more accurately. I also made a major change to handling unique calls and those that did not submit a log. For uniques I do a QRZ lookup to make sure the call exists and entity is correct.
For those that did not submit a log, before I looked for their call in other logs and whatever entity had the most hits I assumed was the correct one. However, for N0N there were 3 guys that logged him and two put “MSH” as the entity and one, AH6KO, got it right with “KS” so my program would say “MSH” was correct even though it wasn’t. Now if there are less than 5 hits with other logs I do a QRZ lookup to get the correct entity.

I also changed the verbiage “The location does not match” to “The entity does not match” for consistency.